2.0

+++ src/Nest/CommonOptions/Stats/RefreshStats.cs
@@ -11,10 +11,11 @@ namespace Nest
 	
 		[JsonProperty(PropertyName = "total")]
 		public long Total { get; set; }
+
 		[JsonProperty(PropertyName = "total_time")]
 		public string TotalTime { get; set; }
 		[JsonProperty(PropertyName = "total_time_in_millis")]
-		public double TotalTimeInMilliseconds { get; set; }
+		public long TotalTimeInMilliseconds { get; set; }
 
 	}
 }
