+++ src/Tests/Framework/MockData/Person.cs
@@ -2,12 +2,13 @@
 using System.Linq;
 using Bogus;
 using Nest;
+using Newtonsoft.Json;
 
 namespace Tests.Framework.MockData
 {
 	public class Person
 	{
-		public int Id { get; set; }
+		public long Id { get; set; }
 		public string FirstName { get; set; }
 		public string LastName { get; set; }
 		public string JobTitle { get; set; }
